+++ src/bsm_api_client/client/_account_methods.py
@@ -23,8 +23,13 @@ class AccountMethodsMixin:
     async def async_get_account_details(self) -> UserResponse:
         """Gets the current user's account details.
 
-        Returns:
-            A `UserResponse` object containing the account details.
+        :returns: A `UserResponse` object containing the account details.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            response = await client.async_get_account_details()
         """
         _LOGGER.debug("Fetching account details from /account")
         response = await self._request(
@@ -35,11 +40,16 @@ class AccountMethodsMixin:
     async def async_update_theme(self, payload: ThemeUpdatePayload) -> BaseApiResponse:
         """Updates the current user's theme.
 
-        Args:
-            payload: A `ThemeUpdatePayload` object containing the new theme.
+        :param payload: A `ThemeUpdatePayload` object containing the new theme.
 
-        Returns:
-            A `BaseApiResponse` object indicating the result of the operation.
+        :returns: A `BaseApiResponse` object indicating the result of the operation.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            payload = ...
+            response = await client.async_update_theme(payload)
         """
         _LOGGER.info("Updating theme to %s", payload.theme)
         response = await self._request(
@@ -55,11 +65,16 @@ class AccountMethodsMixin:
     ) -> BaseApiResponse:
         """Updates the current user's profile.
 
-        Args:
-            payload: A `ProfileUpdatePayload` object containing the new profile data.
+        :param payload: A `ProfileUpdatePayload` object containing the new profile data.
 
-        Returns:
-            A `BaseApiResponse` object indicating the result of the operation.
+        :returns: A `BaseApiResponse` object indicating the result of the operation.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            payload = ...
+            response = await client.async_update_profile(payload)
         """
         _LOGGER.info("Updating profile")
         response = await self._request(
@@ -75,11 +90,16 @@ class AccountMethodsMixin:
     ) -> BaseApiResponse:
         """Changes the current user's password.
 
-        Args:
-            payload: A `ChangePasswordPayload` object.
+        :param payload: A `ChangePasswordPayload` object.
 
-        Returns:
-            A `BaseApiResponse` object indicating the result of the operation.
+        :returns: A `BaseApiResponse` object indicating the result of the operation.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            payload = ...
+            response = await client.async_change_password(payload)
         """
         _LOGGER.info("Changing password")
         response = await self._request(