+++ src/bsm_api_client/client/_users_methods.py
@@ -29,8 +29,13 @@ class UsersMethodsMixin:
     async def async_get_users(self) -> List[UserResponse]:
         """Gets a list of all users.
 
-        Returns:
-            A list of `UserResponse` objects.
+        :returns: A list of `UserResponse` objects.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            response = await client._request()
         """
         _LOGGER.debug("Fetching users from /users/list")
         response = await self._request(
@@ -41,11 +46,15 @@ class UsersMethodsMixin:
     async def async_delete_user(self, user_id: int) -> BaseApiResponse:
         """Deletes a user.
 
-        Args:
-            user_id: The ID of the user to delete.
+        :param user_id: The ID of the user to delete.
 
-        Returns:
-            A `BaseApiResponse` object.
+        :returns: A `BaseApiResponse` object.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            response = await client.async_delete_user()
         """
         _LOGGER.info("Deleting user %s", user_id)
         response = await self._request(
@@ -56,12 +65,16 @@ class UsersMethodsMixin:
     async def async_update_user_role(self, user_id: int, role: str) -> BaseApiResponse:
         """Updates a user's role.
 
-        Args:
-            user_id: The ID of the user.
-            role: The new role.
+        :param user_id: The ID of the user.
+        :param role: The new role.
 
-        Returns:
-            A `BaseApiResponse` object.
+        :returns: A `BaseApiResponse` object.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            response = await client.async_update_user_role()
         """
         _LOGGER.info("Updating role for user %s to %s", user_id, role)
         response = await self._request(
@@ -75,11 +88,15 @@ class UsersMethodsMixin:
     async def async_disable_user(self, user_id: int) -> BaseApiResponse:
         """Disables a user.
 
-        Args:
-            user_id: The ID of the user.
+        :param user_id: The ID of the user.
 
-        Returns:
-            A `BaseApiResponse` object.
+        :returns: A `BaseApiResponse` object.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            response = await client.async_disable_user()
         """
         _LOGGER.info("Disabling user %s", user_id)
         response = await self._request(
@@ -90,11 +107,15 @@ class UsersMethodsMixin:
     async def async_enable_user(self, user_id: int) -> BaseApiResponse:
         """Enables a user.
 
-        Args:
-            user_id: The ID of the user.
+        :param user_id: The ID of the user.
 
-        Returns:
-            A `BaseApiResponse` object.
+        :returns: A `BaseApiResponse` object.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            response = await client.async_enable_user()
         """
         _LOGGER.info("Enabling user %s", user_id)
         response = await self._request(
@@ -105,11 +126,15 @@ class UsersMethodsMixin:
     async def async_generate_invite_token(self, role: str) -> Dict[str, Any]:
         """Generates an invite token.
 
-        Args:
-            role: The role for the new user.
+        :param role: The role for the new user.
 
-        Returns:
-            A dict containing the redirect_url.
+        :returns: A dict containing the redirect_url.
+
+
+        .. rubric:: Example:
+        .. code-block:: python
+
+            response = await client.async_generate_invite_token()
         """
         _LOGGER.info("Generating invite token for role %s", role)
         response = await self._request(